move dialog teachers to teachers.py, and other fixes
