--- docs/source/dialog.rst
@@ -1,12 +1,0 @@
-..
-  Copyright (c) 2017-present, Facebook, Inc.
-  All rights reserved.
-  This source code is licensed under the BSD-style license found in the
-  LICENSE file in the root directory of this source tree. An additional grant
-  of patent rights can be found in the PATENTS file in the same directory.
-
-core.dialog_teacher
-===================================
-.. automodule:: parlai.core.dialog_teacher
-  :members:
-  :exclude-members: __dict__,__weakref__
