--- parlai/core/dialog_teacher.py
@@ -1,385 +1,0 @@
-# Copyright (c) 2017-present, Facebook, Inc.
-# All rights reserved.
-# This source code is licensed under the BSD-style license found in the
-# LICENSE file in the root directory of this source tree. An additional grant
-# of patent rights can be found in the PATENTS file in the same directory.
-
-from .fixed_data_teacher import FixedDataTeacher
-
-from .image_featurizers import ImageLoader
-import sys
-import time
-
-
-class DialogTeacher(FixedDataTeacher):
-    """A base teacher class for doing dialog with fixed chat logs.
-
-    This class provides a set a basic functionality:
-
-    - uses data class to store and query text data
-    - generates action tables to send to the student agent from the data
-    - metrics tracking count of sent vs correctly answered queries
-
-    If you have ``opt.numthreads > 1``, this also activates a shared memory
-    array for the data and lock-protected shared-memory metrics.
-
-    In order to subclass this class, you must implement ``setup_data()`` in your
-    class (or subclass another class which does, like ``FbDialogTeacher``), which
-    reads your data file as an iterator.
-    """
-
-    def __init__(self, opt, shared=None):
-        # Check for setup_data
-        if not hasattr(self, 'setup_data'):
-            raise RuntimeError('Must implement setup_data or subclass a class' +
-                               ' which implements it (e.g. FbDialogTeacher)' +
-                               ' in order to use this class.')
-        super().__init__(opt, shared)
-
-        self.startTime = time.time()
-        self.datatype = opt['datatype']
-        self.training = self.datatype.startswith('train')
-        self.stream = 'stream' in self.datatype.split(':')
-
-        # first initialize any shared objects
-        data_class = StreamDialogData if self.stream else DialogData
-        kwargs = {'cycle': self.training} if self.stream else {}
-        if shared and shared.get('data'):
-            self.data = data_class(opt, shared=shared['data'], **kwargs)
-        else:
-            self.data = data_class(opt, data_loader=self.setup_data,
-                cands=self.label_candidates(), **kwargs)
-
-        self.reset()
-
-    def reset(self):
-        # Reset the dialog so that it is at the start of the epoch,
-        # and all metrics are reset.
-        super().reset()
-        if self.stream:
-            self.data.reset()
-            self.epochDone = False
-
-    def __len__(self):
-        return len(self.data)
-
-    def __iter__(self):
-        self.epochDone = False
-        return self
-
-    def __next__(self):
-        if self.epochDone:
-            raise StopIteration()
-
-    def share(self):
-        shared = super().share()
-        shared['data'] = self.data.share()
-        return shared
-
-    def label_candidates(self):
-        """Returns ``None`` by default, but override this in children (such as
-        ``FbDialogTeacher``) to load up candidate labels for every example.
-        """
-        return None
-
-    def num_episodes(self):
-        return self.data.num_episodes()
-
-    def get(self, episode_idx, entry_idx=0):
-        return self.data.get(episode_idx, entry_idx)[0]
-
-    def next_example(self):
-        if self.stream:
-            action, epoch_done = self.data.get()
-        else:
-            action, epoch_done = super().next_example()
-        return action, epoch_done
-
-
-class DialogData(object):
-    """Provides a data structure for accessing textual dialog data.
-    This can be used whenever the dialog data is a fixed log of chats
-    (i.e not a simulator setting). The logs can include dialog text and possibly
-    supervised labels, candidate labels and rewards.
-
-    All these are stored in this internal data format which is used by the
-    ``DialogTeacher`` class.
-
-    ``data_loader`` is an iterable, with each call returning:
-
-        ``(x, ...), new_episode?``
-
-        Where
-
-        - ``x`` is a query and possibly context
-
-        ``...`` can contain additional fields, specifically
-
-        - ``y`` is an iterable of label(s) for that query
-        - ``r`` is the str reward for getting that query correct
-        - ``c`` is an iterable of label candidates that the student can choose from
-        - ``i`` is a str path to an image on disk, which will be loaded by the data
-          class at request-time. should always point to the raw image file.
-        - ``new_episode?`` is a boolean value specifying whether that example is the start of a new episode. If you don't use episodes set this to ``True`` every time.
-
-
-    ``cands`` can be set to provide a list of candidate labels for every example
-    in this dataset, which the agent can choose from (the correct answer
-    should be in this set).
-
-
-    ``random`` tells the data class whether or not to visit episodes sequentially
-    or randomly when returning examples to the caller.
-    """
-
-    def __init__(self, opt, data_loader=None, cands=None, shared=None, **kwargs):
-        # self.data is a list of episodes
-        # each episode is a tuple of entries
-        # each entry is a tuple of values for the action/observation table
-        if shared:
-            self.image_loader = shared.get('image_loader', None)
-            self.data = shared.get('data', [])
-            self.cands = shared.get('cands', None)
-        else:
-            self.image_loader = ImageLoader(opt)
-            self.data = []
-            self._load(data_loader, opt['datafile'])
-            self.cands = None if cands == None else set(sys.intern(c) for c in cands)
-        self.addedCands = []
-        self.copied_cands = False
-
-    def share(self):
-        shared = {
-            'data': self.data,
-            'cands': self.cands,
-            'image_loader': self.image_loader
-        }
-        return shared
-
-    def __len__(self):
-        """Returns total number of entries available. Each episode has at least
-        one entry, but might have many more.
-        """
-        return sum(len(episode) for episode in self.data)
-
-    def _read_episode(self, data_generator):
-        """Reads one episode at a time from the provided iterator over entries.
-        """
-        episode = []
-        last_cands = None
-        for entry, new in data_generator:
-            if new and len(episode) > 0:
-                yield tuple(episode)
-                episode = []
-                last_cands = None
-
-            # intern all strings so we don't store them more than once
-            new_entry = []
-            if len(entry) > 0:
-                # process text if available
-                if entry[0] is not None:
-                    new_entry.append(sys.intern(entry[0]))
-                else:
-                    new_entry.append(None)
-                if len(entry) > 1:
-                    # process labels if available
-                    if entry[1] is None:
-                        new_entry.append(None)
-                    elif hasattr(entry[1], '__iter__') and type(entry[1]) is not str:
-                        # make sure iterable over labels, not single string
-                        new_entry.append(tuple(sys.intern(e) for e in entry[1]))
-                    else:
-                        raise TypeError('Must provide iterable over labels, not a single string.')
-                    if len(entry) > 2:
-                        # process reward if available
-                        if entry[2] is not None:
-                            new_entry.append(entry[2])
-                        else:
-                            new_entry.append(None)
-                        if len(entry) > 3:
-                            # process label candidates if available
-                            if entry[3] is None:
-                                new_entry.append(None)
-                            elif last_cands and entry[3] is last_cands:
-                                # if cands are shared, say "same" so we
-                                # don't store them again
-                                new_entry.append(
-                                    sys.intern('same as last time'))
-                            elif hasattr(entry[3], '__iter__') and type(entry[3]) is not str:
-                                # make sure iterable over candidates, not single string
-                                last_cands = entry[3]
-                                new_entry.append(tuple(
-                                    sys.intern(e) for e in entry[3]))
-                            else:
-                                raise TypeError('Must provide iterable over label candidates, not a single string.')
-                            if len(entry) > 4 and entry[4] is not None:
-                                new_entry.append(sys.intern(entry[4]))
-
-            episode.append(tuple(new_entry))
-
-        if len(episode) > 0:
-            yield tuple(episode)
-
-    def _load(self, data_loader, datafile):
-        """Loads up data from an iterator over tuples described in the class
-        docs.
-        """
-        for episode in self._read_episode(data_loader(datafile)):
-            self.data.append(episode)
-
-    def num_episodes(self):
-        """Return number of episodes in the dataset."""
-        return len(self.data)
-
-    def get(self, episode_idx, entry_idx=0):
-        """Returns a specific entry from the dataset."""
-        # first look up data
-        episode = self.data[episode_idx]
-        entry = episode[entry_idx]
-        episode_done = entry_idx == len(episode) - 1
-        end_of_data = episode_done and episode_idx == len(self.data) - 1
-
-        # now pack it in a action-observation dictionary
-        table = self.build_table(entry)
-
-        # last entry in this episode
-        table['episode_done'] = episode_done
-        return table, end_of_data
-
-    def build_table(self, entry):
-        """Packs an entry into an action-observation dictionary."""
-        table = {}
-        if entry[0] is not None:
-            table['text'] = entry[0]
-        if len(entry) > 1:
-            if entry[1] is not None:
-                table['labels'] = entry[1]
-            if len(entry) > 2:
-                if entry[2] is not None:
-                    table['reward'] = entry[2]
-                if len(entry) > 3:
-                    if entry[3] is not None:
-                        table['label_candidates'] = entry[3]
-                    if len(entry) > 4 and entry[4] is not None:
-                        img = self.image_loader.load(entry[4])
-                        if img is not None:
-                            table['image'] = img
-
-        if (table.get('labels', None) is not None
-                and self.cands is not None):
-            if self.addedCands:
-                # remove elements in addedCands
-                self.cands.difference_update(self.addedCands)
-                self.addedCands.clear()
-            for label in table['labels']:
-                if label not in self.cands:
-                    # add labels, queue them for removal next time
-                    if not self.copied_cands:
-                        self.cands = self.cands.copy()
-                        self.copied_cands = True
-                    self.cands.add(label)
-                    self.addedCands.append(label)
-            table['label_candidates'] = self.cands
-
-        if 'labels' in table and 'label_candidates' in table:
-            if table['labels'][0] not in table['label_candidates']:
-                raise RuntimeError('true label missing from candidate labels')
-        return table
-
-
-class StreamDialogData(DialogData):
-    """Provides a data structure for streaming textual dialog data.
-    This can be used whenever the dialog data follows the format described in
-    DialogData but cannot fit entirely into memory.
-
-    Additional keyword-argument cycle defines if the stream should restart from
-    the beginning after an epoch is finished (defaults to True).
-    """
-
-    def __init__(self, opt, data_loader=None, cands=None, shared=None, **kwargs):
-        # super() call initiates stream in self.data by calling _load()
-        super().__init__(opt, data_loader, cands, shared, **kwargs)
-        self.cycle = kwargs['cycle'] if 'cycle' in kwargs else True
-        if shared:
-            # auxiliary instances hold pointer to main datastream (in self.data)
-            self.reset_data = shared['reset']
-        else:
-            # main instance holds the stream and shares pointer to it
-            self.data_loader = data_loader
-            self.datafile = opt['datafile']
-            self.reset_data = None
-            self.is_reset = True
-        self.entry_idx = 0
-        self.next_episode = None
-
-    def share(self):
-        shared = super().share()
-        # also share reset method to allow datastream to be reset
-        shared['reset'] = self.reset
-        return shared
-
-    def __len__(self):
-        # unknown
-        return 0
-
-    def _load(self, data_loader, datafile):
-        """Load data generator into data field."""
-        self.data = self._data_generator(data_loader, datafile)
-
-    def _data_generator(self, data_loader, datafile):
-        """Generates data using the iterator over tuples constructed
-        by data_loader.
-        """
-        self.is_reset = False
-        while True:
-            for episode in self._read_episode(data_loader(datafile)):
-                yield episode
-            yield -1
-            while not self.cycle:
-                yield -1
-
-    def num_episodes(self):
-        # unknown
-        return 0
-
-    def get(self):
-        """Returns a the next entry from the stream in the current episode for
-        this instance. When episode is done returns first entry of next episode.
-        """
-        # first look up data
-        if self.next_episode is None:
-            self.next_episode = next(self.data)
-        if self.entry_idx == 0:
-            self.cur_episode = self.next_episode
-            self.next_episode = next(self.data)
-        entry = self.cur_episode[self.entry_idx]
-
-        # now pack it in a action-observation dictionary
-        table = self.build_table(entry)
-
-        episode_done = self.entry_idx == len(self.cur_episode) - 1
-        if episode_done:
-            self.entry_idx = 0
-        else:
-            self.entry_idx += 1
-        end_of_data = episode_done and self.next_episode is -1
-        if end_of_data and self.cycle:
-            self.next_episode = next(self.data)
-
-        # last entry in this episode
-        table['episode_done'] = episode_done
-        return table, end_of_data
-
-    def reset(self):
-        """Reset the datastream to its beginning"""
-        if self.reset_data is not None:
-            # auxiliary instance, reset main datastream
-            self.data = self.reset_data()
-            self.next_episode = None
-        elif not self.is_reset:
-            # if main instance is not reset, reset datastream
-            self._load(self.data_loader, self.datafile)
-            self.is_reset = True
-            self.next_episode = None
-        self.entry_idx = 0
-        return self.data
