+++ parlai/tasks/fromfile/agents.py
@@ -7,16 +7,15 @@
 # This task simply loads the specified file: useful for quick tests without
 # setting up a new task.
 
-from parlai.core.fbdialog_teacher import FbDialogTeacher
+from parlai.core.teachers import FbDialogTeacher
 
 import copy
-import os
 
 class DefaultTeacher(FbDialogTeacher):
     """This task simply loads the specified file: useful for quick tests without
     setting up a new task.
     """
-    
+
     @staticmethod
     def add_cmdline_args(argparser):
         agent = argparser.add_argument_group('FromFile Task Arguments')
